+++ e2e/test-env/src/server/renderer.jsx
@@ -1,16 +1,11 @@
 import React from 'react';
 import { renderToString } from 'react-dom/server';
-import Editor from '../shared/components/Editor';
-import Viewer from '../shared/components/Viewer';
+import RichContentApp from '../../../../examples/main/shared/RichContentApp';
 
 const COMPONENTS = {
   rce: {
-    Components: [Editor],
-    bundleName: 'editor',
-  },
-  combined: {
-    Components: [Editor, Viewer],
-    bundleName: 'combined',
+    Components: [RichContentApp],
+    bundleName: 'index',
   },
 };
 
@@ -18,8 +13,9 @@ export default function renderer() {
   return (req, res) => {
     const [componentId, fixtureName = 'empty'] = req.path.replace(/^\/|\/$/g, '').split('/');
     const isMobile = req.query.mobile === '';
+    const locale = req.query.hebrew === '' ? 'he' : 'en';
     const { Components, bundleName } = COMPONENTS[componentId] || {};
-    const props = { initialState: null, isMobile };
+    const props = { initialState: null, isMobile, locale };
 
     if (!Components) {
       return res.status(404).send(`Component for ${componentId} not found`);
@@ -36,13 +32,14 @@ export default function renderer() {
       html: renderToString(
         <>
           {Components.map((Comp, i) => (
-            <Comp key={i} {...props} />
+            <Comp key={i} mode={'test'} {...props} />
           ))}
         </>
       ),
       initialState: props.initialState,
       bundleName,
       isMobile,
+      locale,
     });
   };
 }