+++ examples/main/shared/editor/Editor.jsx
@@ -1,5 +1,7 @@
 import React, { PureComponent } from 'react';
 import { RichContentEditor, RichContentEditorModal } from 'wix-rich-content-editor';
+import { isSSR } from 'wix-rich-content-common';
+import { convertToRaw } from '@wix/draft-js';
 import * as PropTypes from 'prop-types';
 import ReactModal from 'react-modal';
 import { testImages, testVideos } from './mock';
@@ -24,25 +26,28 @@ export default class Editor extends PureComponent {
   state = {};
   constructor(props) {
     super(props);
-    ReactModal.setAppElement('#root');
+    if (!isSSR()) {
+      ReactModal.setAppElement('#root');
+    }
     this.initEditorProps();
   }
 
   initEditorProps() {
     const mockUpload = (files, updateEntity) => {
-      //mock upload
-      const testItem = testImages[Math.floor(Math.random() * testImages.length)];
-      const data = {
-        id: testItem.photoId,
-        original_file_name: files && files[0] ? files[0].name : testItem.url,
-        file_name: testItem.url,
-        width: testItem.metadata.width,
-        height: testItem.metadata.height,
-      };
-      setTimeout(() => {
-        updateEntity({ data, files });
-        console.log('consumer uploaded', data);
-      }, 500);
+      if (this.props.shouldMockUpload) {
+        const testItem = testImages[Math.floor(Math.random() * testImages.length)];
+        const data = {
+          id: testItem.photoId,
+          original_file_name: files && files[0] ? files[0].name : testItem.url,
+          file_name: testItem.url,
+          width: testItem.metadata.width,
+          height: testItem.metadata.height,
+        };
+        setTimeout(() => {
+          updateEntity({ data, files });
+          console.log('consumer uploaded', data);
+        }, 500);
+      }
     };
     this.helpers = {
       onFilesChange: (files, updateEntity) => mockUpload(files, updateEntity),
@@ -114,6 +119,22 @@ export default class Editor extends PureComponent {
     this.setState({ MobileToolbar, TextToolbar });
   };
 
+  handleChange = editorState => {
+    this.setState({ editorState });
+    if (typeof window !== 'undefined') {
+      // ensures that tests fail when entity map is mutated
+      const raw = convertToRaw(editorState.getCurrentContent());
+      // const raw = deepFreeze(rr);
+      window.__CONTENT_STATE__ = raw;
+      window.__CONTENT_SNAPSHOT__ = {
+        ...raw,
+        // blocks keys are random so for snapshot diffing they are changed to indexes
+        blocks: raw.blocks.map((block, index) => ({ ...block, key: index })),
+      };
+    }
+    this.props.onChange && this.props.onChange(editorState);
+  };
+
   render() {
     const modalStyles = {
       content: Object.assign(
@@ -144,12 +165,12 @@ export default class Editor extends PureComponent {
         )}
         <RichContentEditor
           ref={editor => (this.editor = editor)}
-          onChange={this.props.onChange}
+          onChange={this.handleChange}
           helpers={this.helpers}
           plugins={Plugins.editorPlugins}
           config={Plugins.config}
           editorState={this.props.editorState}
-          // initialState={this.state.initialState}
+          initialState={this.props.initialState}
           isMobile={this.props.isMobile}
           textToolbarType={textToolbarType}
           theme={theme}
@@ -166,13 +187,11 @@ export default class Editor extends PureComponent {
           role="dialog"
           onRequestClose={onRequestClose || this.helpers.closeModal}
         >
-          {this.state.showModal && (
-            <RichContentEditorModal
-              modalsMap={ModalsMap}
-              locale={this.props.locale}
-              {...this.state.modalProps}
-            />
-          )}
+          <RichContentEditorModal
+            modalsMap={ModalsMap}
+            locale={this.props.locale}
+            {...this.state.modalProps}
+          />
         </ReactModal>
       </div>
     );