+++ examples/main/src/ExampleApp.jsx
@@ -3,7 +3,7 @@ import { hot } from 'react-hot-loader/root';
 import React, { PureComponent } from 'react';
 import { ReflexContainer, ReflexElement, ReflexSplitter } from 'react-reflex';
 import { compact, flatMap } from 'lodash';
-import { convertToRaw, createEmpty } from 'wix-rich-content-editor/dist/lib/editorStateConversion';
+import { createEmpty, convertToRaw } from 'wix-rich-content-editor/dist/lib/editorStateConversion';
 import {
   ContentStateEditor,
   ErrorBoundary,
@@ -11,43 +11,36 @@ import {
   SectionHeader,
   SectionContent,
 } from './Components';
-import {
-  generateKey,
-  getStateFromObject,
-  getRequestedLocale,
-  isMobile,
-  loadStateFromStorage,
-  saveStateToStorage,
-} from './utils';
-const Editor = React.lazy(() => import('./editor/Editor'));
-const Viewer = React.lazy(() => import('./viewer/Viewer'));
+import { generateKey, getStateFromObject, loadStateFromStorage, saveStateToStorage } from './utils';
+const Editor = React.lazy(() => import('../shared/editor/Editor'));
+const Viewer = React.lazy(() => import('../shared/viewer/Viewer'));
 const Preview = React.lazy(() => import('./preview/Preview'));
 
-class App extends PureComponent {
+const getContentStateFromEditorState = editorState => convertToRaw(editorState.getCurrentContent());
+
+class ExampleApp extends PureComponent {
   constructor(props) {
     super(props);
-    this.isMobile = isMobile();
     this.state = this.getInitialState();
-    const locale = getRequestedLocale();
-    if (locale !== 'en') {
-      this.setLocale(locale);
-    }
+    disableBrowserBackButton();
   }
 
   getInitialState() {
+    const { isMobile } = this.props;
     const containerKey = generateKey('container');
-    const editorState = createEmpty();
     const localState = loadStateFromStorage();
+    const contentState = getContentStateFromEditorState(createEmpty());
     return {
       containerKey,
-      editorState,
+      contentState,
       isEditorShown: true,
-      isViewerShown: !this.isMobile,
+      isViewerShown: !isMobile,
       isPreviewShown: !this.isMobile,
       isContentStateShown: false,
       viewerResetKey: 0,
       previewResetKey: 0,
       editorResetKey: 0,
+      shouldMockUpload: true,
       ...localState,
     };
   }
@@ -60,16 +53,22 @@ class App extends PureComponent {
     window && window.removeEventListener('resize', this.onContentStateEditorResize);
   }
 
+  onEditorChange = editorState => {
+    this.setState({ contentState: getContentStateFromEditorState(editorState) });
+    this.props.onEditorChange && this.props.onEditorChange(editorState);
+  };
+
   setContentStateEditor = ref => (this.contentStateEditor = ref);
 
   onContentStateEditorChange = obj => {
-    this.setState(getStateFromObject(obj));
+    if (this.props.onEditorChange) {
+      const { editorState } = getStateFromObject(obj);
+      this.props.onEditorChange(editorState);
+    }
   };
 
   onContentStateEditorResize = () =>
     this.contentStateEditor && this.contentStateEditor.refreshLayout();
-
-  onEditorChange = editorState => this.setState({ editorState });
 
   onSectionVisibilityChange = (sectionName, isVisible) => {
     this.setState(
@@ -80,22 +79,14 @@ class App extends PureComponent {
     );
     this.onContentStateEditorResize();
   };
-
-  setLocale = locale => {
-    import(`wix-rich-content-editor/statics/locale/messages_${locale}.json`).then(localeResource =>
-      this.setState({ locale, localeResource: localeResource.default })
-    );
+  onSetLocale = locale => {
+    this.setState({ locale });
+    this.props.setLocale && this.props.setLocale(locale);
   };
 
   renderEditor = () => {
-    const {
-      isEditorShown,
-      editorState,
-      editorIsMobile,
-      staticToolbar,
-      locale,
-      localeResource,
-    } = this.state;
+    const { allLocales, editorState, locale, localeResource, isMobile } = this.props;
+    const { isEditorShown, staticToolbar, shouldMockUpload, editorIsMobile } = this.state;
     const settings = [
       {
         name: 'Mobile',
@@ -106,8 +97,16 @@ class App extends PureComponent {
             editorResetKey: state.editorResetKey + 1,
           })),
       },
+      {
+        name: 'Mock Upload',
+        active: shouldMockUpload,
+        action: () =>
+          this.setState(state => ({
+            shouldMockUpload: !state.shouldMockUpload,
+          })),
+      },
     ];
-    if (!isMobile()) {
+    if (!isMobile) {
       settings.push({
         name: 'Static Toolbar',
         active: staticToolbar,
@@ -116,8 +115,8 @@ class App extends PureComponent {
       settings.push({
         name: 'Locale',
         active: locale,
-        action: selectedLocale => this.setLocale(selectedLocale),
-        items: this.props.allLocales,
+        action: selectedLocale => this.onSetLocale(selectedLocale),
+        items: allLocales,
       });
     }
     return (
@@ -133,7 +132,8 @@ class App extends PureComponent {
               <Editor
                 onChange={this.onEditorChange}
                 editorState={editorState}
-                isMobile={this.state.editorIsMobile || this.isMobile}
+                isMobile={this.state.editorIsMobile || isMobile}
+                shouldMockUpload={this.state.shouldMockUpload}
                 staticToolbar={staticToolbar}
                 locale={locale}
                 localeResource={localeResource}
@@ -181,7 +181,8 @@ class App extends PureComponent {
   };
 
   renderViewer = () => {
-    const { isViewerShown, editorState } = this.state;
+    const { viewerState, isMobile } = this.props;
+    const { isViewerShown } = this.state;
     const settings = [
       {
         name: 'Mobile',
@@ -192,7 +193,6 @@ class App extends PureComponent {
           })),
       },
     ];
-    const viewerState = JSON.parse(JSON.stringify(convertToRaw(editorState.getCurrentContent()))); //emulate initilState passed in by consumers
     return (
       isViewerShown && (
         <ReflexElement key={`viewer-section-${this.state.viewerResetKey}`} className="section">
@@ -203,10 +203,7 @@ class App extends PureComponent {
           />
           <SectionContent>
             <ErrorBoundary>
-              <Viewer
-                initialState={viewerState}
-                isMobile={this.state.viewerIsMobile || this.isMobile}
-              />
+              <Viewer initialState={viewerState} isMobile={this.state.viewerIsMobile || isMobile} />
             </ErrorBoundary>
           </SectionContent>
         </ReflexElement>
@@ -215,7 +212,7 @@ class App extends PureComponent {
   };
 
   renderContentState = () => {
-    const { isContentStateShown, editorState } = this.state;
+    const { contentState, isContentStateShown } = this.state;
     return (
       isContentStateShown && (
         <ReflexElement
@@ -228,7 +225,7 @@ class App extends PureComponent {
             <ContentStateEditor
               ref={this.setContentStateEditor}
               onChange={this.onContentStateEditorChange}
-              contentState={convertToRaw(editorState.getCurrentContent())}
+              contentState={contentState}
             />
           </SectionContent>
         </ReflexElement>
@@ -250,6 +247,7 @@ class App extends PureComponent {
   };
 
   render() {
+    const { isMobile } = this.props;
     const { isEditorShown, isViewerShown, isContentStateShown, isPreviewShown } = this.state;
     const showEmptyState = !isEditorShown && !isViewerShown && !isContentStateShown && !isPreviewShown;
 
@@ -263,7 +261,7 @@ class App extends PureComponent {
           )}
         </ReflexContainer>
         <Fab
-          isMobile={this.isMobile}
+          isMobile={isMobile}
           isEditorShown={isEditorShown}
           isViewerShown={isViewerShown}
           isPreviewShown={isPreviewShown}
@@ -275,4 +273,32 @@ class App extends PureComponent {
   }
 }
 
-export default hot(App);
+function disableBrowserBackButton() {
+  (function(global) {
+    if (typeof global === 'undefined') {
+      throw new Error('window is undefined');
+    }
+
+    var _hash = '!';
+    var noBackPlease = function() {
+      global.location.href += '#';
+
+      // making sure we have the fruit available for juice (^__^)
+      global.setTimeout(function() {
+        global.location.href += '!';
+      }, 50);
+    };
+
+    global.onhashchange = function() {
+      if (global.location.hash !== _hash) {
+        global.location.hash = _hash;
+      }
+    };
+
+    global.onload = function() {
+      noBackPlease();
+    };
+  })(window);
+}
+
+export default hot(ExampleApp);